consolidated to one adminPage file
