+++ src/AdminPage.jsx
@@ -1,101 +1,301 @@
-import { useState, useEffect } from 'react'; 
-import axios from 'axios';
-import NavBar from './components/NavBar.jsx';
-
+import * as React from 'react';
+import PropTypes from 'prop-types';
+import Box from '@mui/material/Box';
 import Table from '@mui/material/Table';
 import TableBody from '@mui/material/TableBody';
-import TableCell from '@mui/material/TableCell';
+import TableCell, { tableCellClasses } from '@mui/material/TableCell';
 import TableContainer from '@mui/material/TableContainer';
 import TableHead from '@mui/material/TableHead';
+import TablePagination from '@mui/material/TablePagination';
 import TableRow from '@mui/material/TableRow';
+import TableSortLabel from '@mui/material/TableSortLabel';
+import Typography from '@mui/material/Typography';
 import Paper from '@mui/material/Paper';
+import { visuallyHidden } from '@mui/utils';
+import axios from 'axios';
+import NavBar from './components/NavBar.jsx';
+import { styled } from '@mui/material/styles';
 
-function AdminPage() {
-    const [cars, setCars] = useState([]);
-
-    const headers = ['Make', 'Model', 'Year', 'Mileage', 'Price', 'Color', 'Fuel', 'Available'];
-
-    const headerTitles = headers.map((header) =>
-        <TableCell key={header}>
-                {header}
-        </TableCell>
-    );
-
-    const changeAvailability = async (id) => {
-        const result = await axios.patch(`http://localhost:3000/availability/${id}`)
-        .then(response => console.log(response.data))
-        .catch(error => console.error(error));
-        return result
-    }
-
-    const handleAvailableClick = (e) => {
-        console.log(e.target)
-        if (e.target.innerHTML === 'Available') {
-            e.target.innerHTML = 'Unavailable'
-        } else {
-            e.target.innerHTML = 'Available'           
-        }
-        changeAvailability(e.target.id)
-    }
-
-    const getCars = async () => {
-        const { data } = await axios.get('http://localhost:3000/cars');
-        return data;
-    }
-
-      useEffect(() => {
-        getCars().then((data) => {
-            setCars(data)
-        })
-    }, [])
-
-    return (
-        <>
-            <NavBar />
-            <TableContainer component={Paper}>
-            <Table sx={{ minWidth: 650 }} aria-label="simple table">
-                <TableHead>
-                    <TableRow>
-                        {headerTitles}
-                    </TableRow>
-                </TableHead>
-                <TableBody>
-                    {cars.map((car) => (
-                        <TableRow
-                            key={Math.random()}
-                            sx={{ '&:last-child td, &:last-child th': { border: 0 } }}
-                        >
-                            <TableCell component="th" scope="row">
-                                {car.make}
-                            </TableCell>
-                            <TableCell component="th" scope="row">
-                                {car.model}
-                            </TableCell>
-                            <TableCell component="th" scope="row">
-                                {car.name.slice(-4)}
-                            </TableCell>
-                            <TableCell component="th" scope="row">
-                                {car.mileage.replace(/\B(?=(\d{3})+(?!\d))/g, ",")}
-                            </TableCell>
-                            <TableCell component="th" scope="row">
-                                ${car.price.replace(/\B(?=(\d{3})+(?!\d))/g, ",")}
-                            </TableCell>
-                            <TableCell component="th" scope="row">
-                                {car.color}
-                            </TableCell>
-                            <TableCell component="th" scope="row">
-                                {car.fuel_type}
-                            </TableCell>
-                            <TableCell component="th" scope="row" id={car.id} onClick={(e) => handleAvailableClick(e)}>
-                                {car.availability ? 'Available' : 'Not Available'}
-                            </TableCell>
-                        </TableRow>
-                    ))}
-                </TableBody>
-            </Table>
-        </TableContainer>
-        </>
-    );
+const getCars = async () => {
+  const { data } = await axios.get('http://localhost:3000/cars');
+  return data;
 }
 
-export default AdminPage;
+const rows = await getCars();
+
+function descendingComparator(a, b, orderBy) {
+  if (b[orderBy] < a[orderBy]) {
+    return -1;
+  }
+  if (b[orderBy] > a[orderBy]) {
+    return 1;
+  }
+  return 0;
+}
+
+function getComparator(order, orderBy) {
+  return order === 'desc'
+    ? (a, b) => descendingComparator(a, b, orderBy)
+    : (a, b) => -descendingComparator(a, b, orderBy);
+}
+
+function stableSort(array, comparator) {
+  const stabilizedThis = array.map((el, index) => [el, index]);
+  stabilizedThis.sort((a, b) => {
+    const order = comparator(a[0], b[0]);
+    if (order !== 0) {
+      return order;
+    }
+    return a[1] - b[1];
+  });
+  return stabilizedThis.map((el) => el[0]);
+}
+
+const headCells = [
+  {
+    id: 'make',
+    numeric: false,
+    disablePadding: false,
+    label: 'Make',
+  },
+  {
+    id: 'model',
+    numeric: false,
+    disablePadding: false,
+    label: 'Model',
+  },
+  {
+    id: 'year',
+    numeric: true,
+    disablePadding: false,
+    label: 'Year',
+  },
+  {
+    id: 'mileage',
+    numeric: true,
+    disablePadding: false,
+    label: 'Mileage',
+  },
+  {
+    id: 'price',
+    numeric: true,
+    disablePadding: false,
+    label: 'Price',
+  },
+  {
+    id: 'color',
+    numeric: false,
+    disablePadding: false,
+    label: 'Color',
+  },
+  {
+    id: 'fuel_type',
+    numeric: false,
+    disablePadding: false,
+    label: 'Fuel',
+  },
+  {
+    id: 'availability',
+    numeric: false,
+    disablePadding: false,
+    label: 'Available',
+  },
+];
+
+const StyledTableCell = styled(TableCell)(({ theme }) => ({
+  [`&.${tableCellClasses.head}`]: {
+    backgroundColor: 'RGB(112,155,203)',
+    color: theme.palette.common.white,
+    fontSize: '20px'
+  },
+}));
+
+function EnhancedTableHead(props) {
+  const { onSelectAllClick, order, orderBy, numSelected, rowCount, onRequestSort } =
+    props;
+  const createSortHandler = (property) => (event) => {
+    onRequestSort(event, property);
+  };
+
+  return (
+    <TableHead>
+      <TableRow>
+        {headCells.map((headCell) => (
+          <StyledTableCell
+            key={headCell.id}
+            align="left"
+            padding={headCell.disablePadding ? 'none' : 'normal'}
+            sortDirection={orderBy === headCell.id ? order : false}
+          >
+            <TableSortLabel
+              active={orderBy === headCell.id}
+              direction={orderBy === headCell.id ? order : 'asc'}
+              onClick={createSortHandler(headCell.id)}
+            >
+              {headCell.label}
+              {orderBy === headCell.id ? (
+                <Box component="span" sx={visuallyHidden}>
+                  {order === 'desc' ? 'sorted descending' : 'sorted ascending'}
+                </Box>
+              ) : null}
+            </TableSortLabel>
+          </StyledTableCell>
+        ))}
+      </TableRow>
+    </TableHead>
+  );
+}
+
+EnhancedTableHead.propTypes = {
+  numSelected: PropTypes.number.isRequired,
+  onRequestSort: PropTypes.func.isRequired,
+  onSelectAllClick: PropTypes.func.isRequired,
+  order: PropTypes.oneOf(['asc', 'desc']).isRequired,
+  orderBy: PropTypes.string.isRequired,
+  rowCount: PropTypes.number.isRequired,
+};
+
+export default function AdminPage() {
+  const [order, setOrder] = React.useState('asc');
+  const [orderBy, setOrderBy] = React.useState('Price');
+  const [selected, setSelected] = React.useState([]);
+  const [page, setPage] = React.useState(0);
+  const [dense, setDense] = React.useState(false);
+  const [rowsPerPage, setRowsPerPage] = React.useState(10);
+
+  const handleRequestSort = (event, property) => {
+    const isAsc = orderBy === property && order === 'asc';
+    setOrder(isAsc ? 'desc' : 'asc');
+    setOrderBy(property);
+  };
+
+  const handleSelectAllClick = (event) => {
+    if (event.target.checked) {
+      const newSelected = rows.map((n) => n.id);
+      setSelected(newSelected);
+      return;
+    }
+    setSelected([]);
+  };
+
+  const handleClick = (event, id) => {
+    const selectedIndex = selected.indexOf(id);
+    let newSelected = [];
+
+    if (selectedIndex === -1) {
+      newSelected = newSelected.concat(selected, id);
+    } else if (selectedIndex === 0) {
+      newSelected = newSelected.concat(selected.slice(1));
+    } else if (selectedIndex === selected.length - 1) {
+      newSelected = newSelected.concat(selected.slice(0, -1));
+    } else if (selectedIndex > 0) {
+      newSelected = newSelected.concat(
+        selected.slice(0, selectedIndex),
+        selected.slice(selectedIndex + 1),
+      );
+    }
+    setSelected(newSelected);
+  };
+
+  const handleChangePage = (event, newPage) => {
+    setPage(newPage);
+  };
+
+  const handleChangeRowsPerPage = (event) => {
+    setRowsPerPage(parseInt(event.target.value, 10));
+    setPage(0);
+  };
+
+  const handleChangeDense = (event) => {
+    setDense(event.target.checked);
+  };
+
+  const isSelected = (id) => selected.indexOf(id) !== -1;
+
+  const emptyRows =
+    page > 0 ? Math.max(0, (1 + page) * rowsPerPage - rows.length) : 0;
+
+  const visibleRows = React.useMemo(
+    () =>
+      stableSort(rows, getComparator(order, orderBy)).slice(
+        page * rowsPerPage,
+        page * rowsPerPage + rowsPerPage,
+      ),
+    [order, orderBy, page, rowsPerPage],
+  );
+
+  return (
+    <>
+      <NavBar/>
+      <Typography variant="h3" component="h3" sx={{ width: '80%', mb:'6', mt: 5, ml:'auto', mr: 'auto'}}>
+        Admin Console:
+      </Typography>
+      <Box sx={{ width: '100%', m: 'auto' }}>
+        <Paper sx={{ width: '80%', mt: 5, m: 'auto' }}>
+          <TableContainer>
+            <Table
+              sx={{ minWidth: 750 }}
+              aria-labelledby="tableTitle"
+              size={dense ? 'small' : 'medium'}
+            >
+              <EnhancedTableHead
+                numSelected={selected.length}
+                order={order}
+                orderBy={orderBy}
+                onSelectAllClick={handleSelectAllClick}
+                onRequestSort={handleRequestSort}
+                rowCount={rows.length}
+              />
+              <TableBody>
+                {visibleRows.map((row, index) => {
+                  const isItemSelected = isSelected(row.id);
+                  const labelId = `enhanced-table-checkbox-${index}`;
+
+                  return (
+                    <TableRow
+                      hover
+                      onClick={(event) => handleClick(event, row.id)}
+                      role="checkbox"
+                      aria-checked={isItemSelected}
+                      tabIndex={-1}
+                      key={row.id}
+                      selected={isItemSelected}
+                      sx={{ cursor: 'pointer' }}
+                    >
+                      <TableCell>{row.make}</TableCell>
+                      <TableCell>{row.model}</TableCell>
+                      <TableCell>{row.year}</TableCell>
+                      <TableCell>{row.mileage.replace(/\B(?=(\d{3})+(?!\d))/g, ",")}</TableCell>
+                      <TableCell>${row.price.replace(/\B(?=(\d{3})+(?!\d))/g, ",")}</TableCell>
+                      <TableCell>{row.color}</TableCell>
+                      <TableCell>{row.fuel_type}</TableCell>
+                      <TableCell>{row.availability ? 'Available' : 'Not Available'}</TableCell>
+                    </TableRow>
+                  );
+                })}
+                {emptyRows > 0 && (
+                  <TableRow
+                    style={{
+                      height: (dense ? 33 : 53) * emptyRows,
+                    }}
+                  >
+                    <TableCell colSpan={6} />
+                  </TableRow>
+                )}
+              </TableBody>
+            </Table>
+          </TableContainer>
+          <TablePagination
+            rowsPerPageOptions={[5, 10, 25]}
+            component="div"
+            count={rows.length}
+            rowsPerPage={rowsPerPage}
+            page={page}
+            onPageChange={handleChangePage}
+            onRowsPerPageChange={handleChangeRowsPerPage}
+          />
+        </Paper>
+      </Box>
+    </>
+  );
+}
